Move upgrade functions tests to a separate file
